backend/requirements.txt
fastapi
uvicorn
mcrcon
python-dotenv
